+++ lib/background_worker/base.rb
@@ -7,14 +7,12 @@ module BackgroundWorker
     attr_accessor :uid, :state
 
     def initialize(options = {})
-      Time.zone = Setting.time_zone
-
       @uid = options[:uid]
 
       # Store state persistently, to enable status checkups & progress reporting
       @state = BackgroundWorker::PersistentState.new(@uid, options.except(:uid))
       log("Created #{self.class}")
-      log("Options are: #{options.pretty_inspect}")
+      log("Options are: #{options.inspect}")
     end
 
     # Report progress...
@@ -79,13 +77,13 @@ module BackgroundWorker
       #
       # It will just call your preferred method in the worker.
       def perform(method_name, options = {})
-        BackgroundWorker.verify_active_connections!
+        BackgroundWorker.verify_active_connections! if BackgroundWorker.config.backgrounded
 
         worker = new(options)
         execution = WorkerExecution.new(worker, method_name, options)
         execution.call
       ensure
-        BackgroundWorker.release_connections!
+        BackgroundWorker.release_connections! if BackgroundWorker.config.backgrounded
       end
     end
   end